Refactored heuristicsModal
