+++ assets/js/layout/gameStudyDropdown.js
@@ -1,13 +1,9 @@
-import { Chart, registerables } from 'https://cdn.jsdelivr.net/npm/chart.js@4.4.2/+esm';
-import { Movetext } from '@chesslablab/jsblab';
 import chessboard from './chessboard.js';
 import heuristicsModal from './heuristicsModal.js';
 import progressModal from './progressModal.js';
 import sanMovesTable from './sanMovesTable.js';
 import * as env from '../../env.js';
 import * as variant from '../../variant.js';
-
-Chart.register(...registerables);
 
 const gameStudyDropdown = document.querySelector('#gameStudyDropdown ul');
 
@@ -93,79 +89,7 @@ gameStudyDropdown.children.item(2).addEventListener('click', async (event) => {
   })
   .then(res => res.json())
   .then(res => {
-    const charts = document.getElementById('charts');
-    while (charts.firstChild) {
-      charts.removeChild(charts.firstChild);
-    }
-    res.names.forEach((item, i) => {
-      const data = res.balance.map((value, index) => value[i]);
-      const allEqual = arr => arr.every(val => val === arr[0]);
-      if (!allEqual(data)) {
-        const div = document.createElement('div');
-        const canvas = document.createElement('canvas');
-        div.appendChild(canvas);
-        charts.appendChild(div);
-        new Chart(canvas, {
-          type: 'line',
-          data: {
-            labels: data,
-            datasets: [{
-              label: item,
-              data: data,
-              borderWidth: 2.25,
-              tension: 0.25,
-              borderColor: '#0a0a0a'
-            }]
-          },
-          options: {
-            animation: false,
-            elements: {
-              point:{
-                radius: 0
-              }
-            },
-            scales: {
-              y: {
-                ticks: {
-                  display: false
-                },
-                grid: {
-                  display: false
-                },
-                border: {
-                  display: false
-                },
-                beginAtZero: true,
-                min: -1.1,
-                max: 1.1
-              },
-              x: {
-                ticks: {
-                  display: false
-                },
-                grid: {
-                  display: false
-                },
-                border: {
-                  display: false
-                }
-              }
-            },
-            plugins: {
-              legend: {
-                position: 'bottom',
-                labels: {
-                  boxWidth: 0,
-                  font: {
-                    size: 16
-                  }
-                }
-              }
-            }
-          }
-        });
-      }
-    });
+    heuristicsModal.domElem(res);
   })
   .catch(error => {
     // TODO
